fix(database): Refactoring some of the database entrypoint code
This code is extremely old and kind of messy, this is just refactoring
it to make some improvements as well as to make documentation better.
