+++ server/cmd/beta.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/monetr/monetr/server/config"
+	"github.com/monetr/monetr/server/database"
 	"github.com/monetr/monetr/server/logging"
 	"github.com/monetr/monetr/server/models"
 	"github.com/monetr/monetr/server/util"
@@ -43,7 +44,7 @@ var (
 			if configFileName := configuration.GetConfigFileName(); configFileName != "" {
 				log.WithField("config", configFileName).Info("config file loaded")
 			}
-			db, err := getDatabase(log, configuration, nil)
+			db, err := database.GetDatabase(log, configuration, nil)
 			if err != nil {
 				log.WithError(err).Fatalf("failed to establish database connection")
 				return err
@@ -86,7 +87,7 @@ var (
 			if configFileName := configuration.GetConfigFileName(); configFileName != "" {
 				log.WithField("config", configFileName).Info("config file loaded")
 			}
-			db, err := getDatabase(log, configuration, nil)
+			db, err := database.GetDatabase(log, configuration, nil)
 			if err != nil {
 				log.WithError(err).Fatalf("failed to establish database connection")
 				return err
